Refactor - Split GroupBy methods
